Got queuing working too! Need to test more thoroughly

+++ libraries/I2S.cpp
@@ -9,7 +9,7 @@ I2S_Tx::I2S_Tx() :
         pio_claim_free_sm_and_add_program(&I2S_Tx_compact_program, &pio, &sm, &offset);
     #endif
 }
-I2S_Tx::I2S_Tx(uint32_t* reservedMem, uint8_t width, uint8_t depth) : 
+I2S_Tx::I2S_Tx(uint32_t* reservedMem, uint32_t width, uint32_t depth) : 
         WS_frame_size(0) {
 
     #if     I2S_TX_PROGRAM == I2S_TX_PROGRAM__NAIVE
@@ -21,7 +21,7 @@ I2S_Tx::I2S_Tx(uint32_t* reservedMem, uint8_t width, uint8_t depth) :
     setReservedMem(reservedMem, width, depth);
 }
 
-void I2S_Tx::setReservedMem(uint32_t* reservedMem, uint8_t width, uint8_t depth) {
+void I2S_Tx::setReservedMem(uint32_t* reservedMem, uint32_t width, uint32_t depth) {
     txPingPong.setReservedSpace(reservedMem, width, depth);
 }
 void I2S_Tx::setDefaultData(uint32_t* defaultData) {
